+++ pkg/aws/v1alpha3/awsmachinedeployment/validate_awsmachinedeployment.go
@@ -111,11 +111,6 @@ func (v *Validator) ValidateCreate(request *admissionv1.AdmissionRequest) (bool,
 		return false, microerror.Maskf(parsingFailedError, "unable to parse awsmachinedeployment: %v", err)
 	}
 
-	err = aws.ValidateNamespace(&awsMachineDeployment)
-	if err != nil {
-		return false, microerror.Mask(err)
-	}
-
 	err = aws.ValidateOrganizationLabelContainsExistingOrganization(context.Background(), v.k8sClient.CtrlClient(), &awsMachineDeployment)
 	if err != nil {
 		return false, microerror.Mask(err)