+++ pkg/aws/v1alpha3/machinedeployment/validate_machinedeployment.go
@@ -58,11 +58,6 @@ func (v *Validator) ValidateCreate(request *admissionv1.AdmissionRequest) (bool,
 		return true, nil
 	}
 
-	err = aws.ValidateNamespace(&machineDeployment)
-	if err != nil {
-		return false, microerror.Mask(err)
-	}
-
 	err = v.ValidateCluster(machineDeployment)
 	if err != nil {
 		return false, microerror.Mask(err)